evaluation/requirements.txt
# Evaluation Framework Requirements
# Note: For full DAM-QA functionality, install from root requirements.txt

# Core dependencies for VQA evaluation
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.20.0

# Computer vision and image processing
pillow>=9.5.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Model-specific dependencies
qwen-vl-utils>=0.0.8 